Move macro into separate file
